Revisione dk.test

+++ org.smeup.sys.il.data.nio/src/org/smeup/sys/il/data/nio/T.java
@@ -4,7 +4,7 @@ public class T {
 
 	public static void main(String[] args) {
 
-		System.out.println(((double)5 % 7));
+		System.out.println((5 % 7));
 	}
 
 }
